Refactor app folder and files structure
