+++ rating/v3/pages/_app.tsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import { AppContextWrapper } from "../components/context/app.context";
+import { AppContextWrapper } from "../components/app-context-wrapper";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
